Configuration creation
A terrible first iteration of the UI which communicates with the server
to set up all of the relevant configuration files. There is no UI to set
up what to include and exclude from the backups yet, but the daemon
already writes the configuration files for them, so I can already
validate that restic is able to init and run a backup based on these
files.