Factor out the backups "overview" UI and add stats
This view of the application will eventually show the backups that have
been configured, along with details about current state, ways to
restore, run, set up and look at scheduling, a button to modify the
configuration and so on.

For now it will show the name, a run button, a progress bar and the size
of the backup.